sentiment/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
transformers==4.37.0
torch==2.1.2
psycopg2-binary==2.9.9
sentencepiece==0.1.99
protobuf==4.25.0
numpy==1.24.3
scipy==1.12.0